# public/swagger_v4.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "https://api.bible.build/v2",
      "description": "Online Staging server for testing"
    },
    {
      "url": "https://api.dbp.dev/v2",
      "description": "Internal Staging server for testing"
    }
  ],
  "info": {
    "title": "Digital Bible Platform [Koinos]",
    "description": "This is a simple API",
    "version": "4.0.0",
    "contact": {
      "name": "API Support",
      "url": "https://dbs.org/contact",
      "email": "jon@dbs.org"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
    "/bibles": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/callback"
          },
          {
            "name": "dam_id",
            "in": "query",
            "description": "Setting this Param will limit the results to the bible matching the `bible_id` or `dam_id` provided.",
            "schema": {
              "type": "string",
              "maxLength": 12,
              "example": "ENGKJV"
            }
          },
          {
            "name": "media",
            "in": "query",
            "description": "Setting this Param will filter the results by the media type provided. The Possible Media Types are available via the `v4_bible_filesets.types` route",
            "schema": {
              "type": "string",
              "example": "audio"
            }
          },
          {
            "name": "language",
            "in": "query",
            "description": "Setting this field equal to an iso639-3 code will filter the resulting filesets to match only those containing that language.",
            "schema": {
              "type": "string",
              "example": "eng",
              "minLength": 3,
              "maxLength": 3
            }
          },
          {
            "name": "full_word",
            "in": "query",
            "description": "Setting this param will filter results to match only those Languages where a name exists. Spanish bibles could be found by setting `full_word` to `Spanish` but We recommend using the Iso Code filter `language` instead of this option where possible as it provides a greater level of precision.",
            "schema": {
              "type": "string",
              "example": "English"
            }
          },
          {
            "name": "language_family_code",
            "in": "query",
            "description": "Setting this field equal to an iso639-3 code will filter the resulting filesets to match only those containing that language and it's dialects.",
            "schema": {
              "type": "string",
              "example": "eng",
              "minLength": 3,
              "maxLength": 3
            }
          },
          {
            "name": "updated",
            "in": "query",
            "description": "Setting this param will filter results only created after a specific date",
            "schema": {
              "type": "string",
              "example": "2018-04-14"
            }
          },
          {
            "name": "organization_id",
            "in": "query",
            "description": "Setting this param will filter results to those were created by the organization specified.",
            "schema": {
              "type": "integer",
              "example": "12",
              "minimum": 1
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "description": "Setting this param will filter results by",
            "schema": {
              "type": "string",
              "example": null
            }
          },
          {
            "name": "country",
            "in": "query",
            "description": "Setting this param will filter results by the Country code given. `AL` would return a couple different languages. Albanian Sign Language, Albanian, ect.",
            "schema": {
              "type": "string",
              "example": "AL"
            }
          },
          {
            "name": "bucket",
            "in": "query",
            "description": "Filter by the providing organization's bucket name. Setting this param will filter results to be only those provided through a specific `bucket_id`",
            "schema": {
              "type": "string",
              "example": "dbs-web"
            }
          },
          {
            "name": "filter_by_fileset",
            "in": "query",
            "description": "This when set to false will return records for all Bibles regardless of whether or not it contains a fileset_id",
            "schema": {
              "type": "boolean",
              "example": false
            }
          }
        ],
        "tags": [
          "Bible"
        ],
        "responses": {
          "200": {
            "description": "An array of Bibles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "data": [
                      {
                        "abbr": "AAKWBT",
                        "name": "Ankave 1990 Edition",
                        "vname": "Xwɨyɨá́ Gorɨxoyá Sɨŋɨṕɨrɨnɨ Yayɨyá Pɨńé",
                        "language": "Ankave",
                        "language_autonym": null,
                        "language_altNames": [],
                        "iso": "aak",
                        "date": 1990,
                        "filesets": [
                          {
                            "id": "AAKWBT",
                            "bucket_id": "dbs-web",
                            "set_type_code": "text_plain",
                            "set_size_code": "P",
                            "set_copyright": null,
                            "ietf_code": null,
                            "meta": []
                          },
                          {
                            "id": "AAKWBT",
                            "bucket_id": "dbs-web",
                            "set_type_code": "text_format",
                            "set_size_code": "NT",
                            "set_copyright": null,
                            "ietf_code": "",
                            "meta": []
                          }
                        ]
                      }
                    ]
                  }
                }
              }
            }
          },
          "403": {
            "description": "Your API key didn't work",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "error": {
                      "message": "No Authentication Provided or invalid Key",
                      "status_code": 403
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/bibles/filesets/{id}/{book_id}/{chapter}": {
      "get": {
        "operationId": "v4_bible_filesets.chapter",
        "summary": "Get text for a chapter",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          },
          {
            "name": "dam_id",
            "in": "query",
            "required": true,
            "description": "This when set to a valid Bible_ID this param will filter records by that Bible ID. For a complete set of bible_ids see the `v4_bible.all` route ",
            "schema": {
              "type": "string",
              "example": ""
            }
          },
          {
            "name": "book_id",
            "in": "query",
            "required": true,
            "description": "This when set to a valid USFM Book ID this param will filter records by that book ID. For a complete set of book_ids see the `v4_bible.allBooks` route ",
            "schema": {
              "type": "string",
              "example": ""
            }
          },
          {
            "name": "chapter_id",
            "in": "query",
            "description": "When set to a chapter number this param will filter records by that chapter number. This number will between 1 & 150.",
            "schema": {
              "type": "integer",
              "example": 3,
              "minimum": 1,
              "maximum": 150
            }
          },
          {
            "name": "verse_start",
            "in": "query",
            "description": "When set to a verse number this param will filter records to those starting at and after that verse_number.",
            "schema": {
              "type": "integer",
              "example": 10,
              "minimum": 1
            }
          },
          {
            "name": "verse_end",
            "in": "query",
            "description": "When set to a verse number this param will filter records to those starting at or before that verse_number.",
            "schema": {
              "type": "integer",
              "example": 15,
              "minimum": 1
            }
          },
          {
            "name": "bucket_id",
            "in": "query",
            "required": true,
            "description": "Specifies the bucket for the resource desired. This field is associated with a particular `fileset_id` in the `v4_bible_filesets.show` & `v4_bible_filesets.index` routes and the complete list of buckets are available via the bucket ",
            "schema": {
              "type": "string",
              "example": "dbp-dev"
            }
          }
        ],
        "tags": [
          "Bible"
        ],
        "responses": {
          "200": {
            "description": "An array of Verse Objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "data": [
                      {
                        "book_id": "GEN",
                        "book_name": "Genesis",
                        "book_name_alt": "Genesis",
                        "chapter": 1,
                        "chapter_alt": "۲",
                        "verse_start": 1,
                        "verse_start_alt": "۲",
                        "verse_end": 1,
                        "verse_end_alt": "۲",
                        "verse_text": "فِي ٱلْبَدْءِ خَلَقَ ٱللهُ ٱلسَّمَاوَاتِ وَٱلْأَرْضَ. "
                      },
                      {
                        "book_id": "GEN",
                        "book_name": "Genesis",
                        "book_name_alt": "Genesis",
                        "chapter": 1,
                        "chapter_alt": "۲",
                        "verse_start": 2,
                        "verse_start_alt": "۳",
                        "verse_end": 2,
                        "verse_end_alt": "۳",
                        "verse_text": "وَكَانَتِ ٱلْأَرْضُ خَرِبَةً وَخَالِيَةً، وَعَلَى وَجْهِ ٱلْغَمْرِ ظُلْمَةٌ، وَرُوحُ ٱللهِ يَرِفُّ عَلَى وَجْهِ ٱلْمِيَاهِ. "
                      },
                      {
                        "book_id": "GEN",
                        "book_name": "Genesis",
                        "book_name_alt": "Genesis",
                        "chapter": 1,
                        "chapter_alt": "۲",
                        "verse_start": 3,
                        "verse_start_alt": "۴",
                        "verse_end": 3,
                        "verse_end_alt": "۴",
                        "verse_text": "وَقَالَ ٱللهُ: «لِيَكُنْ نُورٌ»، فَكَانَ نُورٌ. "
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/bibles/filesets/media/types": {
      "get": {
        "operationId": "v4_bible_filesets.types",
        "summary": "A list of Media types delivered by the API",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Bible"
        ],
        "responses": {
          "200": {
            "description": "An array of media types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "audio_drama": "Dramatized Audio",
                    "audio": "Audio",
                    "text_plain": "Plain Text",
                    "text_formatt": "Formatted Text",
                    "video": "Video",
                    "app": "Application"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/bibles/filesets/{id}/download": {
      "get": {
        "operationId": "v4_bible_filesets.download",
        "summary": "Download either a full zipped audio Bible or a selection",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          },
          {
            "name": "bucket_id",
            "in": "query",
            "required": true,
            "description": "This param states which bucket the resource should be downloaded from.",
            "schema": {
              "type": "string",
              "maxLength": 12,
              "example": "dbs-web"
            }
          }
        ],
        "tags": [
          "Bible"
        ],
        "responses": {
          "200": {
            "description": "An array of Bibles",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "example": "file.zip"
                }
              }
            }
          }
        }
      }
    },
    "/bibles/filesets/{id}/podcast": {
      "get": {
        "operationId": "v4_bible.podcast",
        "summary": "An rss ready format for the audio filesets",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Bible"
        ],
        "responses": {
          "200": {
            "description": "An podcastable ready format for the audio filesets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "parameters": {
      "v": {
        "name": "v",
        "in": "query",
        "description": "The Version Number of the API",
        "required": true,
        "example": 4,
        "schema": {
          "type": "string"
        }
      },
      "key": {
        "name": "key",
        "in": "query",
        "description": "The API key that's been generated for you.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "pretty": {
        "name": "pretty",
        "in": "query",
        "description": "A sugar param, simply will format the call for human eyes",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "version": {
        "name": "key",
        "in": "query",
        "description": "The API key that's been generated for you.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "reply": {
        "name": "reply",
        "in": "query",
        "description": "Specifies the response type requested by the caller. Default is json.",
        "schema": {
          "type": "string",
          "enum": [
            "json",
            "html",
            "jsonp",
            "yaml",
            "xml"
          ]
        }
      },
      "callback": {
        "name": "callback",
        "in": "query",
        "description": "If unused the results are as before. If set to true, the results will be prepended with the array of calling arguments used. This does mean that the results will be one layer deeper as the results will be an array where the first element is the array of input arguments and the second element will be the results as normally returned.",
        "schema": {
          "type": "string"
        }
      },
...
    }
  }
}